backend/services/modbus/requirements.txt
pymodbus==2.5.3
paho-mqtt==2.0.0
requests==2.31.0
